Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    float num1, num2, result;
+    int num1, num2, result;
     char operator;
     scanf("%f", &num1);
     scanf("%f", &num2);
